# java/lance-namespace-apache-client/api/openapi.yaml
openapi: 3.1.1
info:
  description: "This OpenAPI specification is a part of the Lance namespace specification.\
    \ It contains 2 parts:\n\nThe `components/schemas`, `components/responses`, `components/examples`,\
    \ `tags` sections define\nthe request and response shape for each operation in\
    \ a Lance Namespace across all implementations.\nSee https://lancedb.github.io/lance-namespace/spec/operations\
    \ for more details.\n\nThe `servers`, `security`, `paths`, `components/parameters`\
    \ sections are for the \nLance REST Namespace implementation, which defines a\
    \ complete REST server that can work with Lance datasets.\nSee https://lancedb.github.io/lance-namespace/spec/impls/rest\
    \ for more details.\n"
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Lance Namespace Specification
  version: 1.0.0
servers:
- description: Generic server URL with all parts configurable
  url: "{scheme}://{host}:{port}/{basePath}"
  variables:
    scheme:
      default: http
    host:
      default: localhost
    port:
      default: "2333"
    basePath:
      default: ""
- description: Server URL when the port can be inferred from the scheme
  url: "{scheme}://{host}/{basePath}"
  variables:
    scheme:
      default: http
    host:
      default: localhost
    basePath:
      default: ""
security:
- OAuth2: []
- BearerAuth: []
...
paths:
...
  /v1/transaction/{id}/describe:
    parameters:
    - $ref: '#/components/parameters/id'
    - $ref: '#/components/parameters/delimiter'
    post:
      description: Return a detailed information for a given transaction
      operationId: DescribeTransaction
      parameters:
      - $ref: '#/components/parameters/id'
      - $ref: '#/components/parameters/delimiter'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DescribeTransactionRequest'
        required: true
      responses:
        "200":
          $ref: '#/components/responses/DescribeTransactionResponse'
        "400":
          $ref: '#/components/responses/BadRequestErrorResponse'
        "401":
          $ref: '#/components/responses/UnauthorizedErrorResponse'
        "403":
          $ref: '#/components/responses/ForbiddenErrorResponse'
        "404":
          $ref: '#/components/responses/NotFoundErrorResponse'
        "503":
          $ref: '#/components/responses/ServiceUnavailableErrorResponse'
        "5XX":
          $ref: '#/components/responses/ServerErrorResponse'
      summary: Describe information about a transaction
      tags:
      - Transaction
      - Metadata
      x-content-type: application/json
      x-accepts:
      - application/json
  /v1/transaction/{id}/alter:
    parameters:
    - $ref: '#/components/parameters/id'
    - $ref: '#/components/parameters/delimiter'
    post:
      operationId: AlterTransaction
      parameters:
      - $ref: '#/components/parameters/id'
      - $ref: '#/components/parameters/delimiter'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlterTransactionRequest'
        required: true
      responses:
        "200":
          $ref: '#/components/responses/AlterTransactionResponse'
        "400":
          $ref: '#/components/responses/BadRequestErrorResponse'
        "401":
          $ref: '#/components/responses/UnauthorizedErrorResponse'
        "403":
          $ref: '#/components/responses/ForbiddenErrorResponse'
        "404":
          $ref: '#/components/responses/NotFoundErrorResponse'
        "409":
          $ref: '#/components/responses/ConflictErrorResponse'
        "503":
          $ref: '#/components/responses/ServiceUnavailableErrorResponse'
        "5XX":
          $ref: '#/components/responses/ServerErrorResponse'
      summary: Alter information of a transaction.
      tags:
      - Transaction
      - Metadata
      x-content-type: application/json
      x-accepts:
      - application/json
components:
...
  parameters:
    id:
      description: |
        `string identifier` of an object in a namespace, following the Lance Namespace spec.
        When the value is equal to the delimiter, it represents the root namespace.
        For example, `v1/namespace/./list` performs a `ListNamespace` on the root namespace.
      explode: false
      in: path
      name: id
      required: true
      schema:
        type: string
      style: simple
    delimiter:
      description: |
        An optional delimiter of the `string identifier`, following the Lance Namespace spec.
        When not specified, the `.` delimiter must be used.
      explode: true
      in: query
      name: delimiter
      required: false
      schema:
        $ref: '#/components/schemas/PageToken'
      style: form
  responses:
...
    DescribeTransactionResponse:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DescribeTransactionResponse'
      description: Response of GetTransaction
    AlterTransactionResponse:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AlterTransactionResponse'
      description: Response of AlterTransaction
...
    BadRequestErrorResponse:
      content:
        application/json:
          example:
            type: /errors/bad-request
            title: Malformed request
            status: 400
            detail: ""
            instance: /v1/namespaces
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      description: "Indicates a bad request error. It could be caused by an unexpected\
        \ request body format or other forms of request validation failure, such as\
        \ invalid json. Usually serves application/json content, although in some\
        \ cases simple text/plain content might be returned by the server's middleware."
    UnauthorizedErrorResponse:
      content:
        application/json:
          example:
            type: /errors/unauthorized-request
            title: No valid authentication credentials for the operation
            status: 401
            detail: ""
            instance: /v1/namespaces
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      description: Unauthorized. The request lacks valid authentication credentials
        for the operation.
    ForbiddenErrorResponse:
      content:
        application/json:
          example:
            type: /errors/forbidden-request
            title: Not authorized to make this request
            status: 403
            detail: ""
            instance: /v1/namespaces
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      description: Forbidden. Authenticated user does not have the necessary permissions.
    NotFoundErrorResponse:
      content:
        application/json:
          example:
            type: /errors/not-found-error
            title: Not found Error
            status: 404
            detail: ""
            instance: "/v1/namespaces/{ns}"
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      description: A server-side problem that means can not find the specified resource.
...
    ConflictErrorResponse:
      content:
        application/json:
          example:
            type: /errors/conflict
            title: The namespace has been concurrently modified
            status: 409
            detail: ""
            instance: "/v1/namespaces/{ns}"
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      description: The request conflicts with the current state of the target resource.
    ServiceUnavailableErrorResponse:
      content:
        application/json:
          example:
            type: /errors/service-unavailable
            title: Slow down
            status: 503
            detail: ""
            instance: /v1/namespaces
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      description: The service is not ready to handle the request. The client should
        wait and retry. The service may additionally send a Retry-After header to
        indicate when to retry.
    ServerErrorResponse:
      content:
        application/json:
          example:
            type: /errors/server-error
            title: Internal Server Error
            status: 500
            detail: ""
            instance: /v1/namespaces
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      description: A server-side problem that might not be addressable from the client
        side. Used for server 5xx errors without more specific documentation in individual
        routes.
  schemas:
    ErrorResponse:
      description: "JSON error response model based on [RFC-7807](https://datatracker.ietf.org/doc/html/rfc7807)"
      example:
        instance: /login/log/abc123
        detail: Authentication failed due to incorrect username or password
        type: /errors/incorrect-user-pass
        title: Incorrect username or password
        status: 404
      properties:
        type:
          description: a URI identifier that categorizes the error
          example: /errors/incorrect-user-pass
          type: string
        title:
          description: "a brief, human-readable message about the error"
          example: Incorrect username or password
          type: string
        status:
          description: "HTTP response code, (if present) it must match the actual\
            \ HTTP code returned by the service"
          example: 404
          maximum: 600
          minimum: 400
          type: integer
        detail:
          description: a human-readable explanation of the error
          example: Authentication failed due to incorrect username or password
          type: string
        instance:
          description: a URI that identifies the specific occurrence of the error
          example: /login/log/abc123
          type: string
      required:
      - type
...
    PageToken:
      description: |-
        An opaque token that allows pagination for list APIs (e.g. ListNamespaces).
        For an initial client request for a list API, if the server cannot return all items in one response, or if there are more items than the `pageSize` specified in the client request, the server must return a `nextPageToken` in the response indicating there are more results available.
        After the initial request, the value of `nextPageToken` from each response must be used by the client as the `pageToken` parameter value for the next request.
        Clients must interpret either `null`, missing value or empty string value of `nextPageToken` from a server response as the end of the listing results.
      type: string
      nullable: true
...
    TransactionStatus:
      enum:
      - QUEUED
      - RUNNING
      - SUCCEEDED
      - FAILED
      - CANCELED
      type: string
    DescribeTransactionRequest:
      example:
        id: id
      properties:
        id:
          type: string
      required:
      - id
    DescribeTransactionResponse:
      example:
        id: id
        properties:
          key: properties
        status: QUEUED
      properties:
        id:
          type: string
        status:
          $ref: '#/components/schemas/TransactionStatus'
        properties:
          additionalProperties:
            type: string
      required:
      - id
      - status
...
    AlterTransactionRequest:
      description: |
        Alter a transaction with a list of actions.
        The server should either succeed and apply all actions, or fail and apply no action.
      example:
        id: id
        actions:
        - setStatusAction:
            status: QUEUED
          unsetPropertyAction:
            mode: SKIP
            key: key
          setPropertyAction:
            mode: OVERWRITE
            value: value
            key: key
        - setStatusAction:
            status: QUEUED
          unsetPropertyAction:
            mode: SKIP
            key: key
          setPropertyAction:
            mode: OVERWRITE
            value: value
            key: key
      properties:
        id:
          type: string
        actions:
          items:
            $ref: '#/components/schemas/AlterTransactionAction'
          minItems: 1
          type: array
      required:
      - actions
      - id
    AlterTransactionResponse:
      example:
        id: id
        status: QUEUED
      properties:
        id:
          type: string
        status:
          $ref: '#/components/schemas/TransactionStatus'
      required:
      - id
      - status
...